requirements.txt
tensorflow==2.15.0
flask==3.0.0
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
pillow==10.1.0
requests==2.31.0
werkzeug==3.0.1